agregado ModeloPartidaTest.java para empezar los tests
