+++ src/main/java/com/controller/ControllerConsola.java
@@ -1,4 +1,7 @@
-package com.jcg.maven;
+package com.controller;
+
+import com.model.*;
+import com.view.VistaConsola;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
