+++ src/main/java/com/model/GanadorBehavior.java
@@ -1,4 +1,4 @@
-package com.jcg.maven;
+package com.model;
 
 import java.util.ArrayList;
