+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,4 +1,4 @@
 Manifest-Version: 1.0
 Created-By: Apache Maven 3.6.1
-Main-Class: com.jcg.maven.App
+Main-Class: com.App
